+++ src/App.scss
@@ -129,18 +129,6 @@
   font-weight: bold;
 }
 
-.header {
-  width: 100%;
-
-  .linkedin {
-    width: 1.5em;
-    background-color: #fff;
-    border-radius: 0.3em;
-    margin: 0.5em;
-    align-self: flex-end;
-  }
-}
-
 .download-icon {
   position: fixed;
   bottom: 1.4em;
