--- src/components/Header.js
@@ -1,21 +1,0 @@
-import React from 'react';
-
-function Header() {
-  return (
-    <div className="header">
-      <a
-        href="https://www.linkedin.com/in/luis-angel-espa%C3%B1a-yepez-a384471a5/"
-        target={'_blank'}
-        rel="noreferrer"
-      >
-        <img
-          src="https://cdn-icons-png.flaticon.com/512/174/174857.png"
-          alt=""
-          className="linkedin"
-        />
-      </a>
-    </div>
-  );
-}
-
-export default Header;
