+++ src/components/Linkedin/Linkedin.scss
@@ -1,0 +1,7 @@
+.linkedin {
+  width: 1.5em;
+  background-color: #fff;
+  border-radius: 0.3em;
+  margin: 0 0.5em;
+  align-self: flex-end;
+}
