+++ src/components/Logo/Logo.js
@@ -1,0 +1,36 @@
+import React from 'react';
+
+export default function Logo({className}) {
+    return (
+        <div className={className}>
+            <svg width="90"
+                 height="101"
+                 viewBox="0 0 102 101"
+                 fill="none"
+                 xmlns="http://www.w3.org/2000/svg"
+            >
+                <path
+                    d="M39 1.60008C19.5 6.80008 4.7 22.6001 1.5 41.7001L0.800003 46.0001H22C41.8 46.0001 43.5 46.1001 46.6 48.1001C48.5 49.2001 50 50.6001 50 51.1001C50 52.9001 51.9 52.0001 52.5 49.9001C53.7 46.2001 60 40.1001 63.8 39.0001C65.9 38.5001 70.1 38.0001 73.3 38.0001H79.1L75.5 34.9001C68.5 28.7001 67.3 22.0001 72.8 19.9001C74.8 19.1001 76.3 19.2001 78.8 20.4001C81.8 21.9001 82.4 21.9001 85.1 20.5001C86.8 19.6001 89.1 19.2001 90.3 19.5001C94.1 20.6001 80.6 8.50008 74.9 5.70008C63.4 0.100076 50.2 -1.39992 39 1.60008Z"
+                    fill="#A52E94"
+                />
+                <path
+                    d="M94.5 26.9C94.1 28.3 91.4 31.6 88.6 34.4L83.5 39.4L82.9 54.4C82.1 73.8 80.3 79.1 70.9 89.4C66 94.8 60.7 98.3 55 100.1C49.9 101.6 58.2 101.1 64.5 99.5C73.6 97.1 81.4 92.4 88 85.3C97.7 75 102 64.1 102 50.5C101.9 42.2 100.9 37.1 97.7 29.8C95.5 24.9 95.3 24.7 94.5 26.9Z"
+                    fill="#A52E94"
+                />
+                <path
+                    d="M63.4 45C58.8 46.8 55.2 55.1 54.9 64.5C54.7 72.6 50.5 79.2 42.9 83C37.9 85.6 28.7 85.6 21.2 83L15.5 81.1L20.1 85.3C36 100 57.8 98.9 69.5 83C75.3 75.1 77 68.9 77 55.5V44L71.3 44.1C68.1 44.1 64.5 44.5 63.4 45Z"
+                    fill="#A52E94"
+                />
+                <path
+                    d="M5 53.5001V56.0001H16.9C26.2 56.0001 29 56.3001 30 57.5001C31 58.7001 31 59.3001 30 60.5001C29 61.7001 26.3 62.0001 17.8 62.0001C8.2 62.0001 6.9 62.2001 7.4 63.6001C7.7 64.5001 8 65.4001 8 65.6001C8 65.8001 12.7 66.0001 18.5 66.0001C27.6 66.0001 29.1 66.2001 29.9 67.8001C31.6 70.9001 29.3 72.0001 21.2 72.0001C12.5 72.0001 12.5 73.2001 21.3 77.1001C31.6 81.5001 42.8 79.7001 47 72.7001C51.1 66.0001 48.5 54.3001 42.4 52.0001C41 51.4001 32 51.0001 22.4 51.0001H5V53.5001Z"
+                    fill="#A52E94"
+                />
+                <path
+                    d="M82.0723 27.3578C72.3069 16.6285 65.4136 24.2923 82.0723 34C94.1355 27.8687 96.4332 14.5846 82.0723 27.3578Z"
+                    fill="#A52E94"
+                />
+                <circle cx="71.5" cy="50.2432" r="3.5" fill="white"/>
+            </svg>
+        </div>
+    );
+}
